requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pypdf
python-docx
numpy
scikit-learn
openai
sentence-transformers
